# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntakeIn": {
        "additionalProperties": false,
        "properties": {
          "ai_screening_enabled": {
            "default": false,
            "title": "Ai Screening Enabled",
            "type": "boolean"
          },
          "approval_mode": {
            "const": "two_level",
            "default": "two_level",
            "title": "Approval Mode",
            "type": "string"
          },
          "blind_review": {
            "default": true,
            "title": "Blind Review",
            "type": "boolean"
          },
          "closes_at": {
            "format": "date-time",
            "title": "Closes At",
            "type": "string"
          },
          "cohort_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cohort Id"
          },
          "description": {
            "default": "",
            "maxLength": 4000,
            "title": "Description",
            "type": "string"
          },
          "eligibility_rules": {
            "items": {
              "$ref": "#/components/schemas/EligibilityRule"
            },
            "maxItems": 20,
            "title": "Eligibility Rules",
            "type": "array"
          },
          "min_reviews": {
            "default": 1,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Min Reviews",
            "type": "integer"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "opens_at": {
            "format": "date-time",
            "title": "Opens At",
            "type": "string"
          },
          "program_id": {
            "format": "uuid",
            "title": "Program Id",
            "type": "string"
          },
          "quota": {
            "exclusiveMinimum": 0.0,
            "maximum": 100000.0,
            "title": "Quota",
            "type": "integer"
          },
          "rounds": {
            "items": {
              "$ref": "#/components/schemas/RoundIn"
            },
            "maxItems": 6,
            "minItems": 1,
            "title": "Rounds",
            "type": "array"
          },
          "triage_config": {
            "additionalProperties": true,
            "title": "Triage Config",
            "type": "object"
          }
        },
        "required": [
          "program_id",
          "name",
          "opens_at",
          "closes_at",
          "quota",
          "rounds"
        ],
        "title": "IntakeIn",
        "type": "object"
      },
      "IntakeOut": {
        "properties": {
          "ai_screening_enabled": {
            "title": "Ai Screening Enabled",
            "type": "boolean"
          },
          "approval_mode": {
            "title": "Approval Mode",
            "type": "string"
          },
          "blind_review": {
            "title": "Blind Review",
            "type": "boolean"
          },
          "closes_at": {
            "format": "date-time",
            "title": "Closes At",
            "type": "string"
          },
          "cohort_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cohort Id"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "eligibility_rules": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Eligibility Rules",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "min_reviews": {
            "title": "Min Reviews",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "opens_at": {
            "format": "date-time",
            "title": "Opens At",
            "type": "string"
          },
          "program_id": {
            "format": "uuid",
            "title": "Program Id",
            "type": "string"
          },
          "quota": {
            "title": "Quota",
            "type": "integer"
          },
          "rounds": {
            "items": {
              "$ref": "#/components/schemas/RoundOut"
            },
            "title": "Rounds",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "program_id",
          "cohort_id",
          "name",
          "description",
          "opens_at",
          "closes_at",
          "quota",
          "status",
          "rounds",
          "approval_mode",
          "ai_screening_enabled",
          "blind_review",
          "min_reviews",
          "eligibility_rules",
          "counts"
        ],
        "title": "IntakeOut",
        "type": "object"
      },
      "IntakePatch": {
        "additionalProperties": false,
        "properties": {
          "ai_screening_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Screening Enabled"
          },
          "blind_review": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blind Review"
          },
          "closes_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closes At"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "eligibility_rules": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EligibilityRule"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eligibility Rules"
          },
          "min_reviews": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Reviews"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "quota": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota"
          },
          "triage_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triage Config"
          }
        },
        "title": "IntakePatch",
        "type": "object"
      },
...
      "MovedOut": {
        "properties": {
          "moved": {
            "title": "Moved",
            "type": "integer"
          }
        },
        "required": [
          "moved"
        ],
        "title": "MovedOut",
        "type": "object"
      },
...
      "ProgramCreatedOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Name",
            "type": "object"
          }
        },
        "required": [
          "id",
          "code",
          "name"
        ],
        "title": "ProgramCreatedOut",
        "type": "object"
      },
      "ProgramIn": {
        "properties": {
          "code": {
            "pattern": "^[a-z][a-z0-9_-]{1,39}$",
            "title": "Code",
            "type": "string"
          },
          "name_en": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En"
          },
          "name_vi": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name Vi",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name_vi"
        ],
        "title": "ProgramIn",
        "type": "object"
      },
      "ProgramOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "cohorts": {
            "items": {
              "$ref": "#/components/schemas/CohortListItem"
            },
            "title": "Cohorts",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Name",
            "type": "object"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "cohorts"
        ],
        "title": "ProgramOut",
        "type": "object"
      },
...
      "RoundOut": {
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "type": {
            "default": "review",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label"
        ],
        "title": "RoundOut",
        "type": "object"
      },
      "RubricIn": {
        "additionalProperties": false,
        "properties": {
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/Criterion"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Criteria",
            "type": "array"
          }
        },
        "required": [
          "criteria"
        ],
        "title": "RubricIn",
        "type": "object"
      },
      "RubricOut": {
        "properties": {
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/CriterionOut"
            },
            "title": "Criteria",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "criteria"
        ],
        "title": "RubricOut",
        "type": "object"
      },
      "RubricSavedOut": {
        "properties": {
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/CriterionOut"
            },
            "title": "Criteria",
            "type": "array"
          },
          "round": {
            "title": "Round",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "criteria",
          "round"
        ],
        "title": "RubricSavedOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Talent Hub API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/intakes": {
      "get": {
        "operationId": "list_intakes_api_v1_intakes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IntakeOut"
                  },
                  "title": "Response List Intakes Api V1 Intakes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Intakes",
        "tags": [
          "intakes"
        ]
      },
      "post": {
        "operationId": "create_intake_api_v1_intakes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntakeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Intake",
        "tags": [
          "intakes"
        ]
      }
    },
    "/api/v1/intakes/{intake_id}": {
      "get": {
        "operationId": "get_intake_api_v1_intakes__intake_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Intake",
        "tags": [
          "intakes"
        ]
      },
      "patch": {
        "operationId": "patch_intake_api_v1_intakes__intake_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntakePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Intake",
        "tags": [
          "intakes"
        ]
      }
    },
    "/api/v1/intakes/{intake_id}/close": {
      "post": {
        "operationId": "close_intake_api_v1_intakes__intake_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close Intake",
        "tags": [
          "intakes"
        ]
      }
    },
    "/api/v1/intakes/{intake_id}/publish": {
      "post": {
        "operationId": "publish_intake_api_v1_intakes__intake_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Intake",
        "tags": [
          "intakes"
        ]
      }
    },
    "/api/v1/intakes/{intake_id}/rubrics": {
      "get": {
        "operationId": "get_rubrics_api_v1_intakes__intake_id__rubrics_get",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/RubricOut"
                  },
                  "title": "Response Get Rubrics Api V1 Intakes  Intake Id  Rubrics Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rubrics",
        "tags": [
          "intakes"
        ]
      }
    },
    "/api/v1/intakes/{intake_id}/rubrics/{round_key}": {
      "put": {
        "operationId": "put_rubric_api_v1_intakes__intake_id__rubrics__round_key__put",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "round_key",
            "required": true,
            "schema": {
              "title": "Round Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RubricIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RubricSavedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Rubric",
        "tags": [
          "intakes"
        ]
      }
    },
    "/api/v1/intakes/{intake_id}/start": {
      "post": {
        "operationId": "start_round_api_v1_intakes__intake_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Round",
        "tags": [
          "intakes"
        ]
      }
    },
...
    "/api/v1/programs": {
      "get": {
        "operationId": "list_programs_api_v1_programs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProgramOut"
                  },
                  "title": "Response List Programs Api V1 Programs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Programs",
        "tags": [
          "intakes"
        ]
      },
      "post": {
        "operationId": "create_program_api_v1_programs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramCreatedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Program",
        "tags": [
          "intakes"
        ]
      }
    },
    "/api/v1/programs/{program_id}/cohorts": {
      "post": {
        "operationId": "create_cohort_api_v1_programs__program_id__cohorts_post",
        "parameters": [
          {
            "in": "path",
            "name": "program_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Program Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CohortIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CohortCreatedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Cohort",
        "tags": [
          "intakes"
        ]
      }
    },
...
  }
}